Refactor coupler and add vorticity-mixed boundary condition
Flatten source/coupler/core/ into the package root and split the
monolithic solver into boundary, blending, conservation, interpolation,
remesh, pressure_reference, checkpoint and reporting modules.

Add vpm_bc_mode="vorticity_mixed": the coupler prescribes only the
normal velocity and the VPM normal derivative of the tangential
velocity, instead of constraining all three velocity components. Full
Dirichlet data does not enforce vorticity continuity across the
interface, so a tangential mismatch is free to act as a spurious
boundary vorticity source. An exact-trace replay of the reference on
the coupling faces reproduces the centreline profile to 0.8% but still
fails to separate off-axis beside the body (minimum +0.125 against the
reference's -0.071), and refining the mesh does not recover it.

Also:
- warn when the mesher snaps an STL outward, which silently inflated
  the reference cube to side 1.02273 (+4.60% frontal area) and biased
  every force coefficient reported against the nominal area;
- size both cube meshes so the surface lands on a cell boundary and
  the two cases match level for level;
- sample the reference coupling-box faces every step;
- add the VPM isolation matrix and cube sequence harnesses;
- track docs/references/*.pdf through LFS by directory, not by name.

+++ source/solvers/FVM/assemble/matrix_assembly.py
@@ -542,7 +542,9 @@ def assemble_rhs_from_fluxes_vectorized(flux_data, mesh_data, *, backend: str = 
     if backend != "numpy":
         raise ValueError(f"Unknown operator backend {backend!r}")
 
-    b = np.bincount(owners, weights=-flux_vf, minlength=n_elements)
+    # ``np.bincount`` returns an integer array when the interior-face slice is
+    # empty, even though boundary contributions below are floating point.
+    b = np.bincount(owners, weights=-flux_vf, minlength=n_elements).astype(np.float64, copy=False)
     b += np.bincount(neighbours, weights=flux_vf, minlength=n_elements)
 
     # Boundary faces
